- Ongoing IdGenerator refactoring

+++ ums-api/src/main/java/org/ums/domain/model/mutable/MutableApplicationCCI.java
@@ -12,7 +12,7 @@ import org.ums.enums.ApplicationType;
  * Created by My Pc on 7/11/2016.
  */
 public interface MutableApplicationCCI extends ApplicationCCI, Mutable, MutableLastModifier,
-    MutableIdentifier<Integer> {
+    MutableIdentifier<Long> {
   void setSemester(final Semester pSemester);
 
   void setSemesterId(final Integer pSemesterId);
